fix: - improve the user process and visualization on front-end.
- update projects to NET 10.0.

- create Kurochou.Data project for repositories.
