+++ Kurochou.DI/Kurochou.DI.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
+    <TargetFramework>net10.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
@@ -19,6 +19,7 @@
 	  <ProjectReference Include="..\Kurochou.App\Kurochou.App.csproj" />
 	  <ProjectReference Include="..\Kurochou.Domain\Kurochou.Domain.csproj" />
 	  <ProjectReference Include="..\Kurochou.Infra\Kurochou.Infra.csproj" />
+	  <ProjectReference Include="..\Kurochou.Data\Kurochou.Data.csproj" />
 	</ItemGroup>
 
 </Project>
