+++ Kurochou.Data/Kurochou.Data.csproj
@@ -7,12 +7,13 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.Google" Version="8.0.25" />
+    <ProjectReference Include="..\Kurochou.App\Kurochou.App.csproj" />
+    <ProjectReference Include="..\Kurochou.Domain\Kurochou.Domain.csproj" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Kurochou.Domain\Kurochou.Domain.csproj" />
-    <ProjectReference Include="..\Kurochou.Infra\Kurochou.Infra.csproj" />
+  	<PackageReference Include="Dapper" Version="2.1.66" />
+  	<PackageReference Include="Npgsql" Version="8.0.7" />
   </ItemGroup>
 
 </Project>
