+++ Kurochou.Data/Repositories/ClipRepository.cs
@@ -2,7 +2,7 @@
 using Kurochou.Domain.Interface.Repository;
 using System.Data;
 
-namespace Kurochou.Infra.Repositories;
+namespace Kurochou.Data.Repositories;
 
 public class ClipRepository(IDbConnection conn) : Repository<Clip>(conn), IClipRepository
 {
