+++ Kurochou.Domain/DTO/UserWithClipCount.cs
@@ -1,0 +1,12 @@
+﻿using Kurochou.Domain.Enum;
+
+namespace Kurochou.Domain.DTO;
+
+public class UserWithClipCount
+{
+    public Guid Id { get; set; }
+    public string Username { get; set; } = string.Empty;
+    public UserRole Role { get; set; }
+    public DateTime CreatedAt { get; set; }
+    public int Uploads { get; set; }
+}
