+++ Kurochou.Domain/Interface/Repository/IUserRepository.cs
@@ -1,3 +1,4 @@
+using Kurochou.Domain.DTO;
 using Kurochou.Domain.Entities;
 
 namespace Kurochou.Domain.Interface.Repository;
@@ -6,5 +7,5 @@ public interface IUserRepository : IRepository<User>
 {
     Task<User?> GetByUsernameAsync(string username, CancellationToken cancellationToken);
     Task<User?> GetByGoogleIdAsync(string googleId, CancellationToken cancellation);
-    Task<IEnumerable<User?>> GetUsersAsync(string? username, CancellationToken cancellationToken);
+    Task<IEnumerable<UserWithClipCount?>> GetUsersAsync(string? username, CancellationToken cancellationToken);
 }
